Allow delegating Azure signature to a separate process
The process is accessed over HTTP and the URL to access it can be
configured with the signature_delegation_url parameter in the Azure
disk configuration.

ClickHouse will make a POST request with a json body:
{
    "stringToSing": "..."
}
And expects a JSON response with status 200 and this structure:
{
    "signature": 01234567890abcdefg..."
}

The string to sign matches the format defined by Azure for
signatures.

The canonical request also contains all the required  information
(path, operation...) to let the proxy decide of the request is
allowed and can be signed. It's also enough information to know
the cost of the request.

This commit was applied from the patch file 0067-Delegated_signature_azure.patch

Co-authored-by: Kevin Michel <kevin.michel@aiven.io>
Signed-off-by: Mikhail Koviazin <mikhail.koviazin@aiven.io>